style: 파일 네임 변경 및 Numbers 파일 내 함수 구조체 파일로 이동
