--- RockPaperScissors/RockPaperScissors/Numbers.swift
@@ -1,10 +1,0 @@
-//
-//  Numbers.swift
-//  RockPaperScissors
-//
-//  Created by 김동용 on 2022/04/19.
-//
-
-enum Numbers {
-    static let range = 1...3
-}
